docs: Turn docs into static site using Jekyll
The site can be viewed locally, if you `gem install jekyll` and then
run `jekyll serve docs/`. (You'll need to edit the first key in
docs/_config.yml for the theme to work locally).

And it can be uploaded into GitHub Pages. The repository will need
to be configured to build GitHub Pages content from the `docs/`
subdirectory of the 'master' branch.